backend/pyproject.toml
[tool.poetry]
name = "zapgastos-backend"
version = "0.1.0"
description = "ZapGastos FastAPI Backend - Personal Finance Management System"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.116.1"
uvicorn = {extras = ["standard"], version = "0.35.0"}
sqlalchemy = "2.0.37"
psycopg2 = "2.9.10"
python-dotenv = "1.1.1"
pydantic = "2.11.7"
pydantic-settings = "2.5.2"
requests = "2.32.3"
starlette = "0.47.1"
alembic = "1.13.2"
python-dateutil = "^2.9.0.post0"
google = "^3.0.0"
google-auth = "^2.40.3"
google-auth-oauthlib = "^1.2.2"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.183.0"
python-jose = "^3.5.0"
passlib = "^1.7.4"
slowapi = "^0.1.9"
python-multipart = "^0.0.20"
bcrypt = "^5.0.0"
pgvector = "^0.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"
httpx = "^0.25.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
httpx = "^0.25.2"
factory-boy = "^3.3.0"

[tool.poetry.scripts]
dev = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
start = "uvicorn main:app --host 0.0.0.0 --port 8000"
migrate = "alembic upgrade head"
makemigrations = "alembic revision --autogenerate"
test = "pytest"
test-cov = "pytest --cov=app --cov-report=html"
lint = "flake8 app/"
format = "black app/ && isort app/"
format-check = "black --check app/ && isort --check-only app/"
type-check = "mypy app/"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  migrations
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
skip = ["migrations"]

# mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "psycopg2.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_missing_imports = true

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
